split dns_parser test
